build fix
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@42064 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/VBox/vmm/pdmnetshaper.h
@@ -70,7 +70,7 @@ typedef struct PDMNETSHAPER *PPDMNETSHAPER;
  * @param   pFilter         Pointer to the filter that allocates bandwidth.
  * @param   cbTransfer      Number of bytes to allocate.
  */
-VMMR0DECL(bool) PDMR0NsAllocateBandwidth(PPDMNSFILTER pFilter, uint32_t cbTransfer);
+VMMR0DECL(bool) PDMR0NsAllocateBandwidth(PPDMNSFILTER pFilter, size_t cbTransfer);
 
 /**
  * Obtain bandwidth in a bandwidth group.
@@ -79,7 +79,7 @@ VMMR0DECL(bool) PDMR0NsAllocateBandwidth(PPDMNSFILTER pFilter, uint32_t cbTransf
  * @param   pFilter         Pointer to the filter that allocates bandwidth.
  * @param   cbTransfer      Number of bytes to allocate.
  */
-VMMR3DECL(bool) PDMR3NsAllocateBandwidth(PPDMNSFILTER pFilter, uint32_t cbTransfer);
+VMMR3DECL(bool) PDMR3NsAllocateBandwidth(PPDMNSFILTER pFilter, size_t cbTransfer);
 
 /**
  * Attach network filter driver from bandwidth group.
